feat: add a form context to manage form state
